requirements.txt
flask
torch
numpy
gunicorn
ultralytics
pillow
gdown